initial build for count up timer
